backend/requirements.txt
flask
flask-sqlalchemy
flask_cors
psycopg2
mobilitydb-sqlalchemy
trajectory
pygeoif
pytz
